Move ConfigModule
