+++ common/src/main/java/com/shzhangji/vault/common/ConfigModule.java
@@ -1,4 +1,4 @@
-package com.shzhangji.vault.cdc.inject;
+package com.shzhangji.vault.common;
 
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
