+++ etl/pom.xml
@@ -10,4 +10,12 @@
 
   <modelVersion>4.0.0</modelVersion>
   <artifactId>vault-etl</artifactId>
+
+  <dependencies>
+    <dependency>
+      <groupId>com.shzhangji.vault</groupId>
+      <artifactId>vault-common</artifactId>
+      <version>0.0.1-SNAPSHOT</version>
+    </dependency>
+  </dependencies>
 </project>
