+++ pom.xml
@@ -48,6 +48,12 @@
         <version>1.18.24</version>
         <scope>provided</scope>
       </dependency>
+      <dependency>
+        <groupId>org.slf4j</groupId>
+        <artifactId>slf4j-api</artifactId>
+        <version>1.7.36</version>
+        <scope>provided</scope>
+      </dependency>
     </dependencies>
   </dependencyManagement>
